requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.2; python_version >= '3.7'
cfgv==3.3.1; python_full_version >= '3.6.1'
distlib==0.3.6
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
django-js-asset==2.0.0; python_version >= '3.6'
django-mptt==0.14.0
django==4.1.2
filelock==3.8.0; python_version >= '3.7'
gunicorn==20.1.0
identify==2.5.8; python_version >= '3.7'
nodeenv==1.7.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
platformdirs==2.5.2; python_version >= '3.7'
pre-commit==2.20.0
psycopg2-binary==2.9.5
pyyaml==6.0; python_version >= '3.6'
setuptools==65.5.0; python_version >= '3.7'
sqlparse==0.4.3; python_version >= '3.5'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
virtualenv==20.16.6; python_version >= '3.6'
